07/05/22

+++ ft_strlcat.c
@@ -6,18 +6,18 @@
 /*   By: fruiz-ca <fruiz-ca@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/26 12:25:18 by fruiz-ca          #+#    #+#             */
-/*   Updated: 2022/04/28 09:06:56 by fruiz-ca         ###   ########.fr       */
+/*   Updated: 2022/05/07 10:23:16 by fruiz-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
-	size_t			dest_len;
-	size_t			src_len;
+	size_t	i;
+	size_t	j;
+	size_t	dest_len;
+	size_t	src_len;
 
 	dest_len = ft_strlen(dest);
 	src_len = ft_strlen(src);
@@ -26,7 +26,7 @@ unsigned int	ft_strlcat(char *dest, const char *src, size_t size)
 	if (dest_len >= size)
 		dest_len = size;
 	if (size > 0 && dest_len < size - 1)
-	{	
+	{
 		while (src[i] != '\0' && (dest_len + i) < (size - 1))
 		{
 			dest[j] = src[i];
